Rewrite MY::postable to compile .elc files on demand, etc.
With an .elc file entry .elc files can be compiled automatically
when .el files are updated. Perhaps useful when developing. The
command for compiling allows for elisp compilation to fail.

The Makefile now uses the 'install' program instead of cp.

The Makefile honours DESTDIR being set, which is necessary for
packaging modules.